+++ src/main/java/dev/majek/pc/mechanic/PlayerQuit.java
@@ -26,10 +26,13 @@ package dev.majek.pc.mechanic;
 import dev.majek.pc.PartyChat;
 import dev.majek.pc.command.party.PartyLeave;
 import dev.majek.pc.data.Restrictions;
+import dev.majek.pc.data.object.Party;
 import dev.majek.pc.data.object.User;
 import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.player.PlayerQuitEvent;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Handles the player quit event.
@@ -38,14 +41,44 @@ public class PlayerQuit extends Mechanic {
 
   @EventHandler
   public void onPlayerQuit(PlayerQuitEvent event) {
-    User user = PartyChat.dataHandler().getUser(event.getPlayer());
-    if (user.isInParty() && !PartyChat.dataHandler().getConfigBoolean(PartyChat.dataHandler().mainConfig, "persistent-parties"))
+    PartyChat.debug(null, "Calling #onPlayerQuit");
+    User user = PartyChat.dataHandler().getUser(event.getPlayer().getUniqueId());
+    if (user == null) {
+      PartyChat.debug(null, "null user, returning...");
+      return;
+    }
+
+    if (
+        user.isInParty() && user.getParty() != null && user.isLeader()
+            && !PartyChat.dataHandler().persistentParties
+            && PartyChat.dataHandler().getConfigBoolean(PartyChat.dataHandler().mainConfig, "disband-on-leader-leave")
+    ) {
+      disbandParty(user.getParty());
+    }
+    else if (user.isInParty() && !PartyChat.dataHandler().getConfigBoolean(PartyChat.dataHandler().mainConfig, "persistent-parties")) {
       PartyLeave.execute(user, null, true);
+      PartyChat.debug(null, "Executing PartyLeave#execute");
+    }
 
     // Remove the player from the party if they left the server due to a ban
     Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(PartyChat.core(), () -> {
       if (Restrictions.isBanned(event.getPlayer()))
         PartyChat.commandHandler().getCommand("leave").execute(user.getPlayer(), new String[0], true);
     }, 20L); // Give hooked plugins time to figure their shit out
+  }
+
+  private void disbandParty(@NotNull Party party) {
+    // Disband the party
+    for (User user : party.getMembers()) {
+      user.setPartyID(null);
+      user.setInParty(false);
+      user.setPartyOnly(false);
+      user.setPartyChatToggle(false);
+      Player member = user.getPlayer();
+      if (member != null && member.isOnline())
+        PartyChat.messageHandler().sendMessageWithReplacement(member, "party-disbanded", "%partyName%", party.getName());
+    }
+    PartyChat.partyHandler().deleteParty(party);
+    PartyChat.partyHandler().removeFromPartyMap(party.getId());
   }
 }